Updated the project jsp page to show tasks for a given project in column format as Open, In Process, and Completed.  Also added the ability to delete a task from the project (deletes the task from the task table and the user task table).
The ability to add tasks to a given project is also implemented - tasks are added to the Task database table and the UserTask table for the user who created the task.  

The drag and drop functionality is included using sortable columns, but I still need to research how to update task status (in the database) on drag and drop actions.    

Also added a css file, 2 new servlets for adding/deleting tasks, and a couple updates to the DAO.